pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "b39tools"
version = "0.1.0"
description = "Tools for working with BIP-39 assets."
readme = "README.md"
license = { text = "MIT" }
dependencies = [
  "age",
  "base58",
  "mnemonic",
  "pycoin",
  "pysha3",
  "shamir-mnemonic",
  'importlib-metadata; python_version<"3.8"',
]

[project.scripts]
create-bip39-account = "b39tools.accountcreator:main"
